pyproject.toml
[tool.poetry]
name = "ai"
version = "0.1.0"
description = ""
authors = ["AroobMushtaq <aroobmushtaq786@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12"
fastapi = { extras = ["standard"], version = ">=0.115.12,<0.116.0" }
uvicorn = ">=0.34.2,<0.35.0"
pymongo = "^4.15.4"
uv = ">=0.7.20,<0.8.0"
sqlalchemy = ">=2.0.41,<3.0.0"
asyncpg = ">=0.30.0,<0.31.0"
jwt = ">=1.4.0,<2.0.0"
numpy = ">=2.3.2,<3.0.0"
matplotlib = ">=3.10.5,<4.0.0"
pandas = ">=2.3.1,<3.0.0"
opencv-python = "<4.12"
motor = "^3.7.1"
python-dotenv = "^1.2.1"
certifi = "^2025.11.12"
google-generativeai = "^0.8.5"
openai = "^2.8.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
